refactor workspace doctor categories
