+++ .sampo/changesets/672-split-workspace-doctor.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Refactor the workspace doctor into focused package, block, binding, and feature modules while preserving the existing `wp-typia doctor` output and adding direct coverage for the extracted binding diagnostics.
