+++ packages/wp-typia-project-tools/src/runtime/cli-doctor-workspace-bindings.ts
@@ -1,0 +1,229 @@
+import fs from "node:fs";
+import path from "node:path";
+
+import { parseScaffoldBlockMetadata } from "@wp-typia/block-runtime/blocks";
+
+import {
+	checkExistingFiles,
+	createDoctorCheck,
+	WORKSPACE_BINDING_EDITOR_ASSET,
+	WORKSPACE_BINDING_EDITOR_SCRIPT,
+	WORKSPACE_BINDING_SERVER_GLOB,
+} from "./cli-doctor-workspace-shared.js";
+import { escapeRegex } from "./php-utils.js";
+
+import type { DoctorCheck } from "./cli-doctor.js";
+import type { WorkspaceInventory } from "./workspace-inventory.js";
+import type { WorkspaceProject } from "./workspace-project.js";
+
+function checkWorkspaceBindingBootstrap(projectDir: string, packageName: string): DoctorCheck {
+	const packageBaseName = packageName.split("/").pop() ?? packageName;
+	const bootstrapPath = path.join(projectDir, `${packageBaseName}.php`);
+	if (!fs.existsSync(bootstrapPath)) {
+		return createDoctorCheck(
+			"Binding bootstrap",
+			"fail",
+			`Missing ${path.basename(bootstrapPath)}`,
+		);
+	}
+
+	const source = fs.readFileSync(bootstrapPath, "utf8");
+	const hasServerGlob = source.includes(WORKSPACE_BINDING_SERVER_GLOB);
+	const hasEditorEnqueueHook = source.includes("enqueue_block_editor_assets");
+	const hasEditorScript = source.includes(WORKSPACE_BINDING_EDITOR_SCRIPT);
+	const hasEditorAsset = source.includes(WORKSPACE_BINDING_EDITOR_ASSET);
+
+	return createDoctorCheck(
+		"Binding bootstrap",
+		hasServerGlob && hasEditorEnqueueHook && hasEditorScript && hasEditorAsset ? "pass" : "fail",
+		hasServerGlob && hasEditorEnqueueHook && hasEditorScript && hasEditorAsset
+			? "Binding source PHP and editor bootstrap hooks are present"
+			: "Missing binding source PHP require glob or editor enqueue hook",
+	);
+}
+
+function checkWorkspaceBindingSourcesIndex(
+	projectDir: string,
+	bindingSources: WorkspaceInventory["bindingSources"],
+): DoctorCheck {
+	const indexRelativePath = [path.join("src", "bindings", "index.ts"), path.join("src", "bindings", "index.js")].find(
+		(relativePath) => fs.existsSync(path.join(projectDir, relativePath)),
+	);
+
+	if (!indexRelativePath) {
+		return createDoctorCheck(
+			"Binding sources index",
+			"fail",
+			"Missing src/bindings/index.ts or src/bindings/index.js",
+		);
+	}
+
+	const indexPath = path.join(projectDir, indexRelativePath);
+	const source = fs.readFileSync(indexPath, "utf8");
+	const missingImports = bindingSources.filter(
+		(bindingSource) => !source.includes(`./${bindingSource.slug}/editor`),
+	);
+
+	return createDoctorCheck(
+		"Binding sources index",
+		missingImports.length === 0 ? "pass" : "fail",
+		missingImports.length === 0
+			? "Binding source editor registrations are aggregated"
+			: `Missing editor imports for: ${missingImports.map((entry) => entry.slug).join(", ")}`,
+	);
+}
+
+function checkWorkspaceBindingTarget(
+	projectDir: string,
+	workspace: WorkspaceProject,
+	registeredBlockSlugs: Set<string>,
+	bindingSource: WorkspaceInventory["bindingSources"][number],
+): DoctorCheck | undefined {
+	const hasBlock = bindingSource.block !== undefined;
+	const hasAttribute = bindingSource.attribute !== undefined;
+	if (!hasBlock && !hasAttribute) {
+		return undefined;
+	}
+	if (!bindingSource.block || !bindingSource.attribute) {
+		return createDoctorCheck(
+			`Binding target ${bindingSource.slug}`,
+			"fail",
+			"Binding target entries must include both block and attribute.",
+		);
+	}
+	if (!registeredBlockSlugs.has(bindingSource.block)) {
+		return createDoctorCheck(
+			`Binding target ${bindingSource.slug}`,
+			"fail",
+			`Binding target references unknown block "${bindingSource.block}".`,
+		);
+	}
+
+	const blockJsonRelativePath = path.join(
+		"src",
+		"blocks",
+		bindingSource.block,
+		"block.json",
+	);
+	const blockJsonPath = path.join(projectDir, blockJsonRelativePath);
+	const issues: string[] = [];
+	try {
+		const blockJson = parseScaffoldBlockMetadata<Record<string, unknown> & { attributes?: unknown }>(
+			JSON.parse(fs.readFileSync(blockJsonPath, "utf8")),
+		);
+		const attributes = blockJson.attributes;
+		if (!attributes || typeof attributes !== "object" || Array.isArray(attributes)) {
+			issues.push(`${blockJsonRelativePath} must define an attributes object`);
+		} else {
+			const attributeConfig = (attributes as Record<string, unknown>)[bindingSource.attribute];
+			if (
+				!attributeConfig ||
+				typeof attributeConfig !== "object" ||
+				Array.isArray(attributeConfig)
+			) {
+				issues.push(
+					`${blockJsonRelativePath} must declare attribute "${bindingSource.attribute}"`,
+				);
+			}
+		}
+	} catch (error) {
+		issues.push(
+			error instanceof Error
+				? `Unable to read ${blockJsonRelativePath}: ${error.message}`
+				: `Unable to read ${blockJsonRelativePath}.`,
+		);
+	}
+
+	const serverPath = path.join(projectDir, bindingSource.serverFile);
+	if (fs.existsSync(serverPath)) {
+		const serverSource = fs.readFileSync(serverPath, "utf8");
+		const supportedAttributesFilter = `block_bindings_supported_attributes_${workspace.workspace.namespace}/${bindingSource.block}`;
+		if (!serverSource.includes(supportedAttributesFilter)) {
+			issues.push(
+				`${bindingSource.serverFile} must register ${supportedAttributesFilter}`,
+			);
+		}
+		if (!new RegExp(`'${escapeRegex(bindingSource.attribute)}'`, "u").test(serverSource)) {
+			issues.push(
+				`${bindingSource.serverFile} must expose attribute "${bindingSource.attribute}"`,
+			);
+		}
+	} else {
+		issues.push(`Missing ${bindingSource.serverFile}`);
+	}
+
+	const editorPath = path.join(projectDir, bindingSource.editorFile);
+	if (fs.existsSync(editorPath)) {
+		const editorSource = fs.readFileSync(editorPath, "utf8");
+		const blockName = `${workspace.workspace.namespace}/${bindingSource.block}`;
+		const bindingSourceTargetMatch = editorSource.match(
+			/export\s+const\s+BINDING_SOURCE_TARGET\s*=\s*\{([\s\S]*?)\}\s+as\s+const\s*;/u,
+		);
+		if (!bindingSourceTargetMatch) {
+			issues.push(`${bindingSource.editorFile} must export BINDING_SOURCE_TARGET`);
+		} else {
+			const targetSource = bindingSourceTargetMatch[1] ?? "";
+			const attributePattern = new RegExp(
+				`\\battribute\\s*:\\s*["']${escapeRegex(bindingSource.attribute)}["']`,
+				"u",
+			);
+			const blockPattern = new RegExp(
+				`\\bblock\\s*:\\s*["']${escapeRegex(blockName)}["']`,
+				"u",
+			);
+			if (!attributePattern.test(targetSource)) {
+				issues.push(
+					`${bindingSource.editorFile} must document target attribute "${bindingSource.attribute}"`,
+				);
+			}
+			if (!blockPattern.test(targetSource)) {
+				issues.push(`${bindingSource.editorFile} must document target block "${blockName}"`);
+			}
+		}
+	} else {
+		issues.push(`Missing ${bindingSource.editorFile}`);
+	}
+
+	return createDoctorCheck(
+		`Binding target ${bindingSource.slug}`,
+		issues.length === 0 ? "pass" : "fail",
+		issues.length === 0
+			? `${bindingSource.block}.${bindingSource.attribute} is declared and supported`
+			: issues.join("; "),
+	);
+}
+
+export function getWorkspaceBindingDoctorChecks(
+	workspace: WorkspaceProject,
+	inventory: WorkspaceInventory,
+): DoctorCheck[] {
+	const checks: DoctorCheck[] = [];
+
+	if (inventory.bindingSources.length > 0) {
+		checks.push(checkWorkspaceBindingBootstrap(workspace.projectDir, workspace.packageName));
+		checks.push(
+			checkWorkspaceBindingSourcesIndex(workspace.projectDir, inventory.bindingSources),
+		);
+	}
+
+	const registeredBlockSlugs = new Set(inventory.blocks.map((block) => block.slug));
+	for (const bindingSource of inventory.bindingSources) {
+		checks.push(
+			checkExistingFiles(workspace.projectDir, `Binding source ${bindingSource.slug}`, [
+				bindingSource.serverFile,
+				bindingSource.editorFile,
+			]),
+		);
+		const bindingTargetCheck = checkWorkspaceBindingTarget(
+			workspace.projectDir,
+			workspace,
+			registeredBlockSlugs,
+			bindingSource,
+		);
+		if (bindingTargetCheck) {
+			checks.push(bindingTargetCheck);
+		}
+	}
+
+	return checks;
+}
